Basic C++ Program that:
i. Prompts the user to input a starting and ending temperature.
ii. Verifies that the starting temperature is less than the ending
temperature and print an error if not.
iii. Outputs the temperature and corresponding velocity in 1 degree
increments.
iv. Provides with a C style comment block with example inputs